Add test for ParseCAN
