+++ src/CMakeLists.txt
@@ -14,6 +14,8 @@ add_subdirectory(ar)
 
 add_subdirectory(ObstacleSim)
 
+add_definitions(-DCHIP_TYPE=CHIP_TYPE_TEMPLATE)
+
 add_executable(Rover
         Rover.cpp
         CommandLineOptions.cpp
@@ -40,5 +42,15 @@ include(Catch)
 target_link_libraries(tests Catch2::Catch2)
 target_link_libraries(tests ${OpenCV_LIBS})
 catch_discover_tests(tests)
+
+add_executable(TestCAN
+        test.cpp
+        CommandLineOptions.cpp
+        Globals.cpp
+        Network.cpp
+        Networking/ParseCAN.cpp
+        HindsightCAN/CANPacket.c
+        HindsightCAN/CANCommon.c
+        HindsightCAN/PortTemplate.c)
 
 set_property(TARGET Rover PROPERTY CXX_STANDARD 14)
